requirements.txt
black==22.12.0
fastapi==0.88.0
fastapi-utils==0.2.1
python-dotenv==0.21.0
psycopg2-binary==2.9.5
uvicorn==0.20.0
sqlalchemy
alembic
aiohttp
httpx
asyncpg